--- packages/react-client/src/forks/ReactFlightClientConfig.dom-node-turbopack-bundled.js
@@ -1,16 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-export * from 'react-client/src/ReactFlightClientStreamConfigNode';
-export * from 'react-client/src/ReactClientConsoleConfigServer';
-export * from 'react-server-dom-turbopack/src/client/ReactFlightClientConfigBundlerTurbopack';
-export * from 'react-server-dom-turbopack/src/client/ReactFlightClientConfigBundlerTurbopackServer';
-export * from 'react-server-dom-turbopack/src/client/ReactFlightClientConfigTargetTurbopackServer';
-export * from 'react-dom-bindings/src/shared/ReactFlightClientConfigDOM';
-export const usedWithSSR = true;
