--- packages/react-server-dom-turbopack/esm/react-server-dom-turbopack-node-loader.production.js
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-export * from '../src/ReactFlightTurbopackNodeLoader.js';
