--- packages/react-server-dom-turbopack/node-register.js
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-module.exports = require('./src/ReactFlightTurbopackNodeRegister');
