--- packages/react-server-dom-turbopack/npm/client.node.unbundled.js
@@ -1,7 +1,0 @@
-'use strict';
-
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-server-dom-turbopack-client.node.unbundled.production.js');
-} else {
-  module.exports = require('./cjs/react-server-dom-turbopack-client.node.unbundled.development.js');
-}
