--- packages/react-server-dom-turbopack/npm/esm/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
