--- packages/react-server-dom-turbopack/npm/node-register.js
@@ -1,3 +1,0 @@
-'use strict';
-
-module.exports = require('./cjs/react-server-dom-turbopack-node-register.js');
