+++ packages/react-server-dom-turbopack/package.json
@@ -16,20 +16,15 @@
     "client.browser.js",
     "client.edge.js",
     "client.node.js",
-    "client.node.unbundled.js",
     "server.js",
     "server.browser.js",
     "server.edge.js",
     "server.node.js",
-    "server.node.unbundled.js",
     "static.js",
     "static.browser.js",
     "static.edge.js",
     "static.node.js",
-    "static.node.unbundled.js",
-    "node-register.js",
-    "cjs/",
-    "esm/"
+    "cjs/"
   ],
   "exports": {
     ".": "./index.js",
@@ -37,11 +32,7 @@
       "workerd": "./client.edge.js",
       "deno": "./client.edge.js",
       "worker": "./client.edge.js",
-      "node": {
-        "turbopack": "./client.node.js",
-        "webpack": "./client.node.js",
-        "default": "./client.node.unbundled.js"
-      },
+      "node": "./client.node.js",
       "edge-light": "./client.edge.js",
       "browser": "./client.browser.js",
       "default": "./client.browser.js"
@@ -49,16 +40,11 @@
     "./client.browser": "./client.browser.js",
     "./client.edge": "./client.edge.js",
     "./client.node": "./client.node.js",
-    "./client.node.unbundled": "./client.node.unbundled.js",
     "./server": {
       "react-server": {
         "workerd": "./server.edge.js",
         "deno": "./server.browser.js",
-        "node": {
-          "turbopack": "./server.node.js",
-          "webpack": "./server.node.js",
-          "default": "./server.node.unbundled.js"
-        },
+        "node": "./server.node.js",
         "edge-light": "./server.edge.js",
         "browser": "./server.browser.js"
       },
@@ -67,16 +53,11 @@
     "./server.browser": "./server.browser.js",
     "./server.edge": "./server.edge.js",
     "./server.node": "./server.node.js",
-    "./server.node.unbundled": "./server.node.unbundled.js",
     "./static": {
       "react-server": {
         "workerd": "./static.edge.js",
         "deno": "./static.browser.js",
-        "node": {
-          "turbopack": "./static.node.js",
-          "webpack": "./static.node.js",
-          "default": "./static.node.unbundled.js"
-        },
+        "node": "./static.node.js",
         "edge-light": "./static.edge.js",
         "browser": "./static.browser.js"
       },
@@ -85,9 +66,6 @@
     "./static.browser": "./static.browser.js",
     "./static.edge": "./static.edge.js",
     "./static.node": "./static.node.js",
-    "./static.node.unbundled": "./static.node.unbundled.js",
-    "./node-loader": "./esm/react-server-dom-turbopack-node-loader.production.js",
-    "./node-register": "./node-register.js",
     "./src/*": "./src/*.js",
     "./package.json": "./package.json"
   },
