+++ packages/react-server-dom-turbopack/src/__tests__/ReactFlightTurbopackDOMBrowser-test.js
@@ -23,7 +23,7 @@ let ReactServerDOMClient;
 let ReactServerScheduler;
 let reactServerAct;
 
-describe('ReactFlightDOMBrowser', () => {
+describe('ReactFlightTurbopackDOMBrowser', () => {
   beforeEach(() => {
     jest.resetModules();
