+++ packages/react-server-dom-turbopack/src/__tests__/ReactFlightTurbopackDOMReply-test.js
@@ -23,7 +23,7 @@ let ReactServerDOMServer;
 let ReactServerDOMClient;
 let ReactServerScheduler;
 
-describe('ReactFlightDOMReply', () => {
+describe('ReactFlightTurbopackDOMReply', () => {
   beforeEach(() => {
     jest.resetModules();
