+++ packages/react-server-dom-turbopack/src/__tests__/ReactFlightTurbopackDOMReplyEdge-test.js
@@ -20,7 +20,7 @@ let turbopackServerMap;
 let ReactServerDOMServer;
 let ReactServerDOMClient;
 
-describe('ReactFlightDOMReply', () => {
+describe('ReactFlightDOMTurbopackReply', () => {
   beforeEach(() => {
     jest.resetModules();
     // Simulate the condition resolution
