+++ scripts/shared/inlinedHostConfigs.js
@@ -105,49 +105,6 @@ module.exports = [
   {
     shortName: 'dom-node-turbopack',
     entryPoints: [
-      'react-server-dom-turbopack/client.node.unbundled',
-      'react-server-dom-turbopack/src/server/react-flight-dom-server.node.unbundled',
-    ],
-    paths: [
-      'react-dom',
-      'react-dom-bindings',
-      'react-dom/client',
-      'react-dom/profiling',
-      'react-dom/server',
-      'react-dom/server.node',
-      'react-dom/static',
-      'react-dom/static.node',
-      'react-dom/src/server/react-dom-server.node',
-      'react-dom/src/server/ReactDOMFizzServerNode.js', // react-dom/server.node
-      'react-dom/src/server/ReactDOMFizzStaticNode.js',
-      'react-dom-bindings/src/server/ReactDOMFlightServerHostDispatcher.js',
-      'react-dom-bindings/src/server/ReactFlightServerConfigDOM.js',
-      'react-dom-bindings/src/shared/ReactFlightClientConfigDOM.js',
-      'react-server-dom-turbopack',
-      'react-server-dom-turbopack/client.node.unbundled',
-      'react-server-dom-turbopack/server',
-      'react-server-dom-turbopack/server.node.unbundled',
-      'react-server-dom-turbopack/static',
-      'react-server-dom-turbopack/static.node.unbundled',
-      'react-server-dom-turbopack/src/client/ReactFlightDOMClientNode.js', // react-server-dom-turbopack/client.node.unbundled
-      'react-server-dom-turbopack/src/client/ReactFlightClientConfigBundlerNode.js',
-      'react-server-dom-turbopack/src/server/react-flight-dom-server.node.unbundled',
-      'react-server-dom-turbopack/src/server/ReactFlightDOMServerNode.js', // react-server-dom-turbopack/src/server/react-flight-dom-server.node.unbundled
-      'react-server-dom-turbopack/node-register',
-      'react-server-dom-turbopack/src/ReactFlightTurbopackNodeRegister.js',
-      'react-devtools',
-      'react-devtools-core',
-      'react-devtools-shell',
-      'react-devtools-shared',
-      'shared/ReactDOMSharedInternals',
-      'react-server/src/ReactFlightServerConfigDebugNode.js',
-    ],
-    isFlowTyped: true,
-    isServerSupported: true,
-  },
-  {
-    shortName: 'dom-node-turbopack-bundled',
-    entryPoints: [
       'react-server-dom-turbopack/client.node',
       'react-server-dom-turbopack/src/server/react-flight-dom-server.node',
     ],
